Renaming helpers -> utilities based on code review feedback
