+++ src/app/containers/MediaPlayer/utilities/getDefaultImage/index.js
