+++ src/app/containers/MediaPlayer/utilities/getDefaultImage/index.test.js
