+++ src/app/containers/MediaPlayer/utilities/getImagePlaceholder/index.js
