+++ src/app/containers/MediaPlayer/utilities/getImagePlaceholder/index.test.js
