+++ src/app/containers/MediaPlayer/utilities/getImagePlaceholderSrcSet/index.js
