+++ src/app/containers/MediaPlayer/utilities/getImagePlaceholderSrcSet/index.test.js
