+++ src/app/containers/MediaPlayer/utilities/metadata/index.js
