+++ src/app/containers/MediaPlayer/utilities/metadata/index.test.js
